+++ src/app/recipes/recipe-detail/recipe-detail.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { DataStorageService } from 'src/app/shared/data-storage.service';
+import { DataStorageService } from 'src/app/services/data-storage.service';
 
 import { Recipe } from '../recipe.model';
 import { RecipeService } from '../recipe.service';
