+++ src/app/recipes/recipes-resolver.service.ts
@@ -6,7 +6,7 @@ import {
 } from '@angular/router';
 
 import { Recipe } from './recipe.model';
-import { DataStorageService } from '../shared/data-storage.service';
+import { DataStorageService } from '../services/data-storage.service';
 import { RecipeService } from './recipe.service';
 
 @Injectable({ providedIn: 'root' })
