+++ src/app/shared/loading-spinner/loading-spinner.component.css
@@ -11,10 +11,10 @@
   width: 51px;
   height: 51px;
   margin: 6px;
-  border: 6px solid #2102cf;
+  border: 6px solid #c3da86;
   border-radius: 50%;
   animation: lds-ring 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite;
-  border-color: #2102cf transparent transparent transparent;
+  border-color: #c3da86 transparent transparent transparent;
 }
 .lds-ring div:nth-child(1) {
   animation-delay: -0.45s;
